requirements.txt
appdirs==1.4.3
beautifulsoup4==4.6.0
bleach==2.0.0
defusedxml==0.5.0
diff-match-patch==20121119
dj-database-url==0.4.2
Django==1.11.16
django-appconf==1.0.2
django-autocomplete-light==3.2.7
django-avatar==4.0.0
django-bootstrap-form==3.2.1
django-braces==1.11.0
django-cors-headers==2.1.0
django-filter==2.0.0
django-fullurl==0.5
git+https://github.com/DemokratieInBewegung/django-mailer.git
django-mathfilters==0.4.0
django-model-utils==3.0.0
django-notifications-hq==1.2
django-rest-framework==0.1.0
django-reversion==3.0.2
django-reversion-compare==0.8.5
django-tag-parser==2.1
django-user-accounts==2.0.1
djangoajax==2.3.7
djangorestframework==3.7.1
dominate==2.1.17
easy-thumbnails==2.3
gunicorn==19.7.1
html2text==2014.12.29
html5lib==0.999999999
jsonfield==2.0.1
lockfile==0.12.2
markdown2==2.3.4
oauthlib==2.0.1
olefile==0.44
packaging==16.8
Pillow==4.0.0
pinax-notifications==4.0.0
psycopg2==2.7.3.2
pyparsing==2.2.0
python3-openid==3.1.0
pytz==2017.2
requests==2.13.0
requests-oauthlib==0.8.0
six==1.10.0
sparkpost==1.3.5
webencodings==0.5.1
wikipedia==1.4.0
